chore: move nodes directory to directly under workflow service directory
