+++ apps/backend/src/services/workflow/nodes/agents/agent/model-providers/openai/openai.provider.ts
