+++ apps/backend/src/services/workflow/nodes/agents/one-inch-agent/v1/tools/oneinch-create-limit-order-tool/index.ts
@@ -2,9 +2,9 @@ import { RunContext, Tool, tool } from "@openai/agents";
 import { name, description, argsSchema } from "./info.js";
 import { Hex } from "viem";
 import { OneInchToolsContext } from "../types.js";
-import { logger } from "../../../../../../../../../utils/logger/index.js";
+import { logger } from "../../../../../../../../utils/logger/index.js";
 import { privateKeyToAccount } from "viem/accounts";
-import { OneInchService } from "../../../../../../../../blockchain/platform/oneinch/index.js";
+import { OneInchService } from "../../../../../../../blockchain/platform/oneinch/index.js";
 
 /**
  * Tool for the agent create limit order using 1Inch.
