+++ apps/backend/src/services/workflow/nodes/agents/one-inch-agent/v1/tools/oneinch-quotation-tool/info.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import { ToolNames } from "../../types.js";
-import { supportedNetworks } from "../../../../../../../../blockchain/platform/oneinch/constants.js";
-import { SupportedNetworks } from "../../../../../../../../blockchain/platform/oneinch/types.js";
+import { supportedNetworks } from "../../../../../../../blockchain/platform/oneinch/constants.js";
+import { SupportedNetworks } from "../../../../../../../blockchain/platform/oneinch/types.js";
 
 const networksList = Array.from(supportedNetworks.keys());
