+++ apps/backend/src/services/workflow/nodes/agents/one-inch-agent/v1/tools/types.ts
@@ -1,0 +1,9 @@
+import { RPCProvider } from "../../../../../../blockchain/rpc/types.js";
+import { SupportedNetworks } from "../../../../../../blockchain/platform/oneinch/types.js";
+
+export type OneInchToolsContext = {
+  oneInchApiKey?: string;
+  rpcProvider?: RPCProvider;
+  evmPrivateKey?: string;
+  enabledNetworks?: SupportedNetworks[];
+};
