+++ apps/backend/src/services/workflow/nodes/triggers/chat/chat.trigger.node.ts
