+++ apps/backend/src/services/workflow/nodes/triggers/chat/v1/chat.v1.trigger.node.ts
